+++ dash_extensions/enrich.py
@@ -37,7 +37,7 @@ from dash import (  # lgtm [py/unused-import]
     page_container,
     page_registry,
     register_page,
-    ctx
+    ctx,
 )
 from dash._callback_context import context_value
 from dash._utils import patch_collections_abc
@@ -62,6 +62,7 @@ DEPENDENCY_APPEND_PREFIX = "dash_extensions_"
 
 
 # region DependencyCollection
+
 
 def build_index(structure, entry, index):
     if isinstance(structure, list):
@@ -161,6 +162,7 @@ class DependencyCollection:
 # endregion
 
 # region Dash blueprint
+
 
 def collect_args(args: Union[Tuple[Any], List[Any]], inputs, outputs):
     for arg in args:
@@ -308,15 +310,17 @@ class DashBlueprint:
         return callbacks, clientside_callbacks
 
     # TODO: Include or not? The plugin still seems a bit immature.
-    def register(self, app: Union[dash.Dash, DashProxy], module, prefix: Union[str, PrefixIdTransform, None]=None, **kwargs):      
+    def register(
+        self, app: Union[dash.Dash, DashProxy], module, prefix: Union[str, PrefixIdTransform, None] = None, **kwargs
+    ):
         # Add prefix transform if supplied.
         if prefix is not None:
             prefix_transform = prefix if isinstance(prefix, PrefixIdTransform) else PrefixIdTransform(prefix)
-            self.transforms.append(prefix_transform)                                          
+            self.transforms.append(prefix_transform)
         # Register the callbacks and page.
         self.register_callbacks(app)
         dash.register_page(module, layout=self._layout_value, **kwargs)
-    
+
     def clear(self):
         self.callbacks = []
         self.clientside_callbacks = []
@@ -359,11 +363,21 @@ class DashProxy(dash.Dash):
     work (e.g. setting a secret key on the server), and exposes convenience functions such as 'hijack'.
     """
 
-    def __init__(self, *args, transforms=None, include_global_callbacks=True, blueprint=None,
-                 prevent_initial_callbacks="initial_duplicate", **kwargs):
+    def __init__(
+        self,
+        *args,
+        transforms=None,
+        include_global_callbacks=True,
+        blueprint=None,
+        prevent_initial_callbacks="initial_duplicate",
+        **kwargs,
+    ):
         super().__init__(*args, prevent_initial_callbacks=prevent_initial_callbacks, **kwargs)
-        self.blueprint = DashBlueprint(transforms,
-                                       include_global_callbacks=include_global_callbacks) if blueprint is None else blueprint
+        self.blueprint = (
+            DashBlueprint(transforms, include_global_callbacks=include_global_callbacks)
+            if blueprint is None
+            else blueprint
+        )
         self.setup_server_lock = threading.Lock()
 
     def callback(self, *args, **kwargs):
@@ -374,7 +388,8 @@ class DashProxy(dash.Dash):
 
     def long_callback(self, *_args, **_kwargs):
         raise NotImplementedError(
-            "The 'long_callback(..)' syntax is not supported, please use 'callback(background=True, ...)' instead.")
+            "The 'long_callback(..)' syntax is not supported, please use 'callback(background=True, ...)' instead."
+        )
 
     def register_celery_tasks(self):
         if sys.argv[0].endswith("celery"):
@@ -455,6 +470,7 @@ def _extract_list_from_kwargs(kwargs: dict, key: str) -> list:
 # endregion
 
 # region Dash transform
+
 
 class DashTransform:
     def __init__(self):
@@ -539,14 +555,16 @@ class BlockingCallbackTransform(StatefulDashTransform):
             end_client_id = f"{callback_id}_end_client"
             start_blocked_id = f"{callback_id}_start_blocked"
             end_blocked_id = f"{callback_id}_end_blocked"
-            self.components.extend([
-                dcc.Store(id=start_client_id),
-                dcc.Store(id=end_server_id),
-                dcc.Store(id=end_client_id),
-                dcc.Store(id=start_blocked_id),
-                dcc.Store(id=start_client_ctx),
-                CycleBreaker(id=end_blocked_id)
-            ])
+            self.components.extend(
+                [
+                    dcc.Store(id=start_client_id),
+                    dcc.Store(id=end_server_id),
+                    dcc.Store(id=end_client_id),
+                    dcc.Store(id=start_blocked_id),
+                    dcc.Store(id=start_client_ctx),
+                    CycleBreaker(id=end_blocked_id),
+                ]
+            )
             # Bind start signal callback.
             start_callback = f"""function()
             {{
@@ -605,7 +623,7 @@ class BlockingCallbackTransform(StatefulDashTransform):
                 end_callback,
                 [Output(end_client_id, "data"), Output(end_blocked_id, "src")],
                 Input(end_server_id, "data"),
-                State(start_blocked_id, "data")
+                State(start_blocked_id, "data"),
             )
             # Modify the original callback to send finished signal.
             num_outputs = len(callback.outputs)
@@ -638,8 +656,8 @@ def skip_input_signal_add_output_signal(num_outputs, out_flex_key, in_flex_key, 
                 outputs = f(*args, **kwargs)
             except Exception:
                 logging.exception(f"Exception raised in blocking callback [{f.__name__}]")
-                outputs = no_update if single_output else [no_update] * num_outputs       
-            
+                outputs = no_update if single_output else [no_update] * num_outputs
+
             return _append_output(outputs, datetime.utcnow().timestamp(), single_output, out_flex_key)
 
         return decorated_function
@@ -653,8 +671,12 @@ def skip_input_signal_add_output_signal(num_outputs, out_flex_key, in_flex_key, 
 
 
 class LogConfig:
-    def __init__(self, log_output, log_writer_map: Dict[int, Callable],
-                 layout_transform: Callable[[List[Component]], List[Component]]):
+    def __init__(
+        self,
+        log_output,
+        log_writer_map: Dict[int, Callable],
+        layout_transform: Callable[[List[Component]], List[Component]],
+    ):
         self.log_output = log_output
         self.log_writer_map = log_writer_map
         self.layout_transform = layout_transform
@@ -670,7 +692,7 @@ def setup_notifications_log_config():
         layout.append(html.Div(id=log_id))
         if dmc.__version__ < "0.14.0":
             layout.append(dmc.NotificationsProvider())
-        layout.append(dmc.NotificationProvider())
+        layout.append(dmc.NotificationProvider(zIndex=2000))
 
         return layout
 
@@ -794,7 +816,73 @@ def bind_logger(logger, single_output, out_flex_key):
 
 # endregion
 
+# region Loading transform
+
+
+class LoadingTransform(DashTransform):
+    """
+    The loading transform makes it easy to trigger a single loading component from multiple callbacks.
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__()
+        kwargs_defaults = {
+            "overlay_style": {
+                "opacity": 0.5,
+                "backgroundColor": "white",
+                "width": "100vw",
+                "height": "100vh",
+                "position": "fixed",
+                "left": 0,
+                "top": 0,
+                "zIndex": 1000,
+                "visibility": "visible",
+            },
+            "delay_show": 500,
+            "delay_hide": 1000,
+            "fullscreen": True,
+            "id": "loading_id",
+        }
+        self.kwargs = {**kwargs_defaults, **kwargs}
+
+    def transform_layout(self, layout):
+        loading = dcc.Loading(html.Div(""), **self.kwargs)
+        layout.children = _as_list(layout.children) + [loading]
+
+    def apply(self, callbacks, clientside_callbacks):
+        callbacks = self.apply_serverside(callbacks)
+        return callbacks, clientside_callbacks
+
+    def apply_serverside(self, callbacks):
+        for callback in callbacks:
+            if not callback.kwargs.get("loading", None):
+                continue
+            # Add the log component as output.
+            single_output = len(callback.outputs) <= 1
+            out_flex_key = callback.outputs.append(Output(self.kwargs["id"], "children", allow_duplicate=True))
+            # Modify the callback function accordingly.
+            f = callback.f
+            callback.f = bind_loading(single_output, out_flex_key)(f)
+
+        return callbacks
+
+
+def bind_loading(single_output, out_flex_key):
+    def wrapper(f):
+        @functools.wraps(f)
+        def decorated_function(*args, **kwargs):
+            outputs = f(*args, **kwargs)
+            return _append_output(outputs, dash.no_update, single_output, out_flex_key)
+
+        return decorated_function
+
+    return wrapper
+
+
+# endregion
+
 # region Cycle breaker transform
+
 
 class CycleBreakerTransform(StatefulDashTransform):
 
@@ -868,24 +956,24 @@ def register(blueprint: DashBlueprint, name: str, prefix=None, **kwargs):
 class PrefixIdTransform(DashTransform):
     def __init__(self, prefix, prefix_func=None, escape=None):
         """
-        The PrefixIdTransform adds a prefix to all component ids of the DashBlueprint, including 
-        their references in callbacks. It is typically used to avoid ID collisions between 
+        The PrefixIdTransform adds a prefix to all component ids of the DashBlueprint, including
+        their references in callbacks. It is typically used to avoid ID collisions between
         blueprints when they are registered on or embedded in the main Dash application.
 
         Args:
             prefix (str): The prefix that will be added to the component_ids.
             prefix_func (callable(str, Component, callable(str)): A function used to modify
-                the ID of the components. This function must accept three arguments: the 
+                the ID of the components. This function must accept three arguments: the
                 prefix string, a Dash component, and the escape function. If not provided,
-                the `prefix_component()` function is used, which relies on `apply_prefix()` 
+                the `prefix_component()` function is used, which relies on `apply_prefix()`
                 to calculate the new ID.
-            escape (callable(str)): A function that will be called by `apply_prefix()` to 
+            escape (callable(str)): A function that will be called by `apply_prefix()` to
                 determine whether the component's ID should remain unaltered (escaped). The
-                function should accept a string (the component_id) and return a boolean. 
-                By default, `default_prefix_escape()` is used, which avoids modifying 
+                function should accept a string (the component_id) and return a boolean.
+                By default, `default_prefix_escape()` is used, which avoids modifying
                 certain IDs that start with "a-" or "anchor-".
 
-        Note: `PrefixIdTransform` is automatically registered as `transforms` by 
+        Note: `PrefixIdTransform` is automatically registered as `transforms` by
                the `DashBlueprint.register()` method when the `prefix` parameter is specified.
         """
         super().__init__()
@@ -1017,7 +1105,7 @@ def filter_args(args_filter):
     def wrapper(f):
         @functools.wraps(f)
         def decorated_function(*args):
-            post_args = list(args[len(args_filter):])
+            post_args = list(args[len(args_filter) :])
             args = list(args[: len(args_filter)])
             filtered_args = [arg for j, arg in enumerate(args) if not args_filter[j]] + post_args
             return f(*filtered_args)
@@ -1036,6 +1124,7 @@ def trigger_filter(args):
 # endregion
 
 # region Multiplexer transform
+
 
 class MultiplexerTransform(DashTransform):
     """
@@ -1075,6 +1164,7 @@ def _output_id_without_wildcards(output: Output) -> str:
 # endregion
 
 # region SerializationTransform
+
 
 class SerializationTransform(DashTransform):
 
@@ -1126,9 +1216,11 @@ class SerializationTransform(DashTransform):
     def sort_key(self):
         return 0
 
+
 # endregion
 
 # region DataclassTransform
+
 
 class DataclassTransform(SerializationTransform):
 
@@ -1218,8 +1310,9 @@ class EnrichedOutput(Output):
     Like a normal Output, includes additional properties related to storing the data.
     """
 
-    def __init__(self, component_id, component_property, allow_duplicate=False, backend=None, session_check=None,
-                 arg_check=True):
+    def __init__(
+        self, component_id, component_property, allow_duplicate=False, backend=None, session_check=None, arg_check=True
+    ):
         super().__init__(component_id, component_property, allow_duplicate)
         self.backend = backend
         self.session_check = session_check
@@ -1229,9 +1322,9 @@ class EnrichedOutput(Output):
 class ServersideOutputTransform(SerializationTransform):
     prefix: str = "SERVERSIDE_"
 
-    def __init__(self,
-                 backends: Optional[List[ServersideBackend]] = None,
-                 default_backend: Optional[ServersideBackend] = None):
+    def __init__(
+        self, backends: Optional[List[ServersideBackend]] = None, default_backend: Optional[ServersideBackend] = None
+    ):
         super().__init__()
         # Per default, use file system backend.
         if backends is None:
@@ -1245,7 +1338,7 @@ class ServersideOutputTransform(SerializationTransform):
             return data
         if not data.startswith(self.prefix):
             return data
-        obj = json.loads(data[len(self.prefix):])
+        obj = json.loads(data[len(self.prefix) :])
         backend = self._backend_registry[obj["backend_uid"]]
         value = backend.get(obj["key"], ignore_expired=True)
         return value
@@ -1328,6 +1421,7 @@ class Dash(DashProxy):
 
 # region Utils
 
+
 def _as_list(item):
     if item is None:
         return []
@@ -1384,5 +1478,6 @@ def _check_multi(item):
 
 def plotly_jsonify(data):
     return json.loads(json.dumps(data, cls=plotly.utils.PlotlyJSONEncoder))
+
 
 # endregion